refactor: move favourite cards to separate elements
