+++ src/entities/offer/ui/offer-card.tsx
@@ -1,4 +1,6 @@
 import {FC, MouseEventHandler} from 'react';
+import {Link} from 'react-router-dom';
+import {RoutePath} from '../../../app/routes.ts';
 import {Offer} from '../model/types.ts';
 
 type OfferCardProps = {
@@ -46,7 +48,7 @@ export const OfferCard: FC<OfferCardProps> = ({offer, onMouseEnter}) => {
           </div>
         </div>
         <h2 className="place-card__name">
-          <a href="#">{title}</a>
+          <Link to={`/${RoutePath.OfferPage}/${offer.id}`}>{title}</Link>
         </h2>
         <p className="place-card__type">{type}</p>
       </div>
