+++ src/pages/favourites-page/cards/favourite-offer-card.tsx
@@ -1,0 +1,50 @@
+import {FC} from 'react';
+import {Link} from 'react-router-dom';
+import {RoutePath} from '../../../app/routes.ts';
+import {Offer} from '../../../entities/offer/model/types.ts';
+
+type FavouriteOfferCardProps = {
+  offer: Offer;
+};
+
+export const FavouriteOfferCard: FC<FavouriteOfferCardProps> = ({offer}) => {
+  return (
+    <article key={offer.id} className="favorites__card place-card">
+      {offer.isPremium && (
+        <div className="place-card__mark">
+          <span>Premium</span>
+        </div>
+      )}
+      <div className="favorites__image-wrapper place-card__image-wrapper">
+        <a href="#">
+          <img className="place-card__image" src={offer.photoUrl} width="150" height="110" alt="Place image"/>
+        </a>
+      </div>
+      <div className="favorites__card-info place-card__info">
+        <div className="place-card__price-wrapper">
+          <div className="place-card__price">
+            <b className="place-card__price-value">&euro;{offer.costPerNight}</b>
+            <span className="place-card__price-text">&#47;&nbsp;night</span>
+          </div>
+          <button className="place-card__bookmark-button place-card__bookmark-button--active button" type="button">
+            <svg className="place-card__bookmark-icon" width="18" height="19">
+              <use xlinkHref="#icon-bookmark"></use>
+            </svg>
+            <span className="visually-hidden">In bookmarks</span>
+          </button>
+        </div>
+        <div className="place-card__rating rating">
+          <div className="place-card__stars rating__stars">
+            <span style={{width: `${20 * offer.stars}%`}}></span>
+            <span className="visually-hidden">Rating</span>
+          </div>
+        </div>
+        <h2 className="place-card__name">
+          <Link to={`/${RoutePath.OfferPage}/${offer.id}`}>{offer.title}</Link>
+        </h2>
+        <p className="place-card__type">{offer.type}</p>
+      </div>
+    </article>
+  );
+};
+
