+++ src/pages/favourites-page/favourites-page.tsx
@@ -1,23 +1,12 @@
 import {FC} from 'react';
 import {Offer} from '../../entities/offer/model/types.ts';
+import {FavouriteOfferCardList} from './cards/favourite-offer-card-list.tsx';
 
 type FavouritesPageProps = {
   offers: Offer[];
 };
 
-type OffersByCity = Record<Offer['city'], Offer[]>;
-
 export const FavouritesPage: FC<FavouritesPageProps> = ({offers}) => {
-
-  const offersByCity: OffersByCity = offers.reduce((result, offer) => {
-    const city = offer.city;
-    if (!result[city]) {
-      result[city] = [];
-    }
-    result[city].push(offer);
-    return result;
-  }, {} as OffersByCity);
-
   return (
     <div className="page">
       <header className="header">
@@ -53,59 +42,7 @@ export const FavouritesPage: FC<FavouritesPageProps> = ({offers}) => {
         <div className="page__favorites-container container">
           <section className="favorites">
             <h1 className="favorites__title">Saved listing</h1>
-            <ul className="favorites__list">
-              {Object.entries(offersByCity).map(([city, cityOffers]) => (
-                <li key={city} className="favorites__locations-items">
-                  <div className="favorites__locations locations locations--current">
-                    <div className="locations__item">
-                      <a className="locations__item-link" href="#">
-                        <span>{city}</span>
-                      </a>
-                    </div>
-                  </div>
-                  <div className="favorites__places">
-                    {cityOffers.map((offer) => (
-                      <article key={offer.id} className="favorites__card place-card">
-                        {offer.isPremium && (
-                          <div className="place-card__mark">
-                            <span>Premium</span>
-                          </div>
-                        )}
-                        <div className="favorites__image-wrapper place-card__image-wrapper">
-                          <a href="#">
-                            <img className="place-card__image" src={offer.photoUrl} width="150" height="110" alt="Place image"/>
-                          </a>
-                        </div>
-                        <div className="favorites__card-info place-card__info">
-                          <div className="place-card__price-wrapper">
-                            <div className="place-card__price">
-                              <b className="place-card__price-value">&euro;{offer.costPerNight}</b>
-                              <span className="place-card__price-text">&#47;&nbsp;night</span>
-                            </div>
-                            <button className="place-card__bookmark-button place-card__bookmark-button--active button" type="button">
-                              <svg className="place-card__bookmark-icon" width="18" height="19">
-                                <use xlinkHref="#icon-bookmark"></use>
-                              </svg>
-                              <span className="visually-hidden">In bookmarks</span>
-                            </button>
-                          </div>
-                          <div className="place-card__rating rating">
-                            <div className="place-card__stars rating__stars">
-                              <span style={{width: `${20 * offer.stars}%`}}></span>
-                              <span className="visually-hidden">Rating</span>
-                            </div>
-                          </div>
-                          <h2 className="place-card__name">
-                            <a href="#">{offer.title}</a>
-                          </h2>
-                          <p className="place-card__type">{offer.type}</p>
-                        </div>
-                      </article>
-                    ))}
-                  </div>
-                </li>
-              ))}
-            </ul>
+            <FavouriteOfferCardList offers={offers}/>
           </section>
         </div>
       </main>
